Revert "easier install with pip install ."
This reverts commit f62b5a70e82c27f3e662ecc82b5c56f9c1db3fb6.
